lib/bootstrap.sh: continue testing when fail to load rootfs partition
Most time rootfs partition is not a must for test, it's ok to move on.
One thing to do is it impacts test like phoronix, which is not
considered in this patch.

Two cases after this patch
1. switch from a general job to phoronix job (need rsync-rootfs) will
fail at the call of rsync_rootf, which impacts current job and phoronix
job
2. local boot switches to phoronix job, it fails at rsync_rootf and
reboot to get next job.

Both cases behaves differently without this patch, which is like
1. run a general job which fails at mount_rootfs, and reboot
to get next job (phoronix)
2. local boot switches to phoronix job and fails at mount_rootfs

Signed-off-by: Philip Li <philip.li@intel.com>