will-it-scale: fix 'affinity'+'threads' paramters
Yu found a issue:

"
Current when running with 2 thread number(1 and 72 for example):
/runtest.py brk1 20 both 1 72
will-it-scale will terminate immediately, because the
the 72 will be treated as the affinity parameter.
"

And Ying pointed that 'threads' is a parameter of variable
length like the "1 72" above, so the 'affnity' parameter
should better be moved before 'threads', and keep 'threads'
always the last one.

Reported-by: Yu Chen <yu.c.chen@intel.com>
Suggested-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Feng Tang <feng.tang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>