repo/linux/ojeda-linux: update branch allowlist and denylist
On 2/11/2022 19:01, Miguel Ojeda wrote:

(By the way, for non-Rust purposes, you are monitoring
https://github.com/ojeda/linux too -- could you please only monitor
the compiler-attributes, clang-format and auxdisplay branches? Those
are the ones I send to linux-next, and sometimes I upload unfinished
work for testing on the GitHub CI, and your bot noticed the temporary
branch and found an issue, but I did not intend for that to happen or
to use resources on your side -- sorry!).

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>