repo/linux/l1k: Mark repository private and enable build reports
I push branches to my repository to allow testing and reviewing before
the patches are sent to the mailing list.  At this stage, I only need
test reports myself.  Others are just irritated if they receive them.
So mark the repository private.

While at it, enable build reports so that I know when 0-day has
completed its tests successfully and the patches can safely be sent
to the mailing list.

Signed-off-by: Lukas Wunner <lukas@wunner.de>