nvdimm: make it possible to specify namespace to reconfig
I would like to use a specified /dev/pmemX as swap device instead of
using all /dev/pmem devices and for this to work, I need to reconfig the
specified namespace to let a single /dev/pmemX appear.

E.g, to use /dev/pmem1 which is on node 1 as swap device, the job file
will have the following:

nvdimm:
  mode: raw
  ns_id: 1
disk:
  nr_pmem: 1
swap:

This has the assumption of all optane DIMMs are configured as dax
devices by default and it *seems* to be the case but I'm not entirely
sure why and whether this will change in the future.

Cc: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Acked-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>