setup/nvdimm: improve single namespace reconfig
It appears the default mode of the namespaces after boot are random and
when the default mode is raw, the current way to make only one specific
namespace as raw will fail.

To make this always work, introduce ns_id as well as ns_mode:

nvdimm:
  mode: dax
  nr_dev: 1
  ns_id: 1
  ns_mode: raw

In the above setup, all namespaces will be firstly configured to dax
mode and then namespace1.0 will be reconfigured as raw mode. So after
the setup, only /dev/pmem1 will appear, no /dev/pmem0.

If ns_id and ns_mode are not specified, the existing behaviour is
retained: all namespaces will be configured to $mode.

"let's leave the echoes there for the sake of manual check"

Cc: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>