distro/depends/blktests-dev: don't install pkg from unstable source
unstable source is a rolling development version of the Debian
distribution containing the latest packages, the unstable repository is
updated every 6 hours.

any pkg/unstable doesn't point to a fix version and would change
frequently which causes unmet dependency issues, so we'd better drop the
unstable source.

== fix issue ==
libtirpc-dev : Depends: libtirpc3 (= 1.3.1-1) but 1.3.2-2 is to be installed

Link: https://wiki.debian.org/DebianUnstable
Signed-off-by: Jie2x Zhou <jie2x.zhou@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>