distro/adaptation/debian-11: add some python pkg mappings
some python2 packages are not available since debian 11, but still there
in debian 10, so we do a mapping to python3 only for debian 11.

== fix issue ==

E: Unable to locate package python-wheel

Package python-docutils is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>