distro/adaptation/debian-11: ignore winpdb
winpdb is only available in debian 10, and it is particularly for
python2. there is no substitution of it in debian 11.

https://github.com/bluebird75/winpdb#readme

The port of winpdb-reborn to Python 3 / WxPython 4 is unfortunately not working very well. So Winpdb for Python 3 does not really work. Feel free to submit Pull Requests to improve the situation. At the moment, this project is stopped.

this package is required by ocfs2test benchmark.

according to the description of winpdb, it is a graphical debugger for
Python. need to monitor whether it will impact ocfs2test after it is
removed.

Link: https://github.com/bluebird75/winpdb
Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>