distro/adaptation/debian-11: add python3-gobject mapping
python3-gobject is not available in debian 11 source.
according to the doc of pygobject, it is replaced by a combination of
multiple packages.

Ubuntu / Debian
Installing the system provided PyGObject:
1.Open a terminal
2.Execute sudo apt install python3-gi python3-gi-cairo gir1.2-gtk-3.0

Fedora
Installing the system provided PyGObject:
Open a terminal
Execute sudo dnf install python3-gobject gtk3

we set different mapping for pygobject on different distros

for debian 10, we stick to use python-gobject because it used to
work well.

for debian 11, we change to the combination of python3-gi,
python3-gi-cairo and gir1.2-gtk-3.0

for fedora, we use python-gobject and gtk3.

== fix issue ==

Package python3-gobject is a virtual package provided by:
  python3-gi 3.14.0-1 [Not candidate version]
E: Package 'python3-gobject' has no installation candidate

Link: https://pygobject.readthedocs.io/en/latest/getting_started.html
Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>