distro/adaptation: add python gobject mapping for ubuntu 20.04 and 21.04
we added a general mapping from python2 to python3 for ubuntu distros
above 20.04, but python gobject package doesn't have a python3 version.

for ubuntu 20.04, old version of python2 gobject is still available, so
we map it back to python-gobject.

for ubuntu 21.04, neither python2 nor python3 gobject is available, so
we ignore ignore this package, it is replaced by a combination of
python3-gi, python3-gi-cairo and gtk3 in new distros.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>