pkg/xfstests/PKGBUILD: disable test in the "dangerous_fuzzers" group
feedback by Dave Chinner for our report "[xfs] 1eb70f54c4: xfstests.xfs.348.fail"
You are running a test in the "dangerous_fuzzers" group that isn't in the auto group.
This is a test for developers to find error detection issues and,
as such, failure is kinda expected. It is not regression test that can be used for pass/fail testing....

Signed-off-by: Jie2x Zhou <jie2x.zhou@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>