bin/lkp-setup-rootfs: support noreboot get next job
For some tests, the submit is broken down into multiple jobs,
but the physical environment required for these jobs is the same,
and for such jobs, restarting the machine every time you finish running
a task can waste a lot of time preparing the machine.
So we need to provide a function that allows users to submit tasks that
do not require a restart of the machine,
so that the test machine can continuously consume jobs
and greatly reduce the cycle of a test round.

Signed-off-by: Wu Zhende <wuzhende666@163.com>