will-it-scale: pass the value of no_affinity to the benchmark
Currently the no_affinity will be set to 1 if the
job.yaml has provided a parameter of no_affinity,
no matter what the value is.

Change it to let no_affinity to specify the value
of no_affinity.

For example, the following parameter in the yaml:

will-it-scale:
    no_affinity: 0

Before this change, no_affinity will be converted to 1
and passed to will-it-scale. After this patch, the
value 0 instead of 1 will be passed to will-it-scale.

Besides that, remove the comment of no_affinity so that
splitting yaml is aware of the no_affinity parameter per
Bebei's suggestion.

Fixes: cf79c4fe43e00a (will-it-scale: update package and add affinity parameter)
Signed-off-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>