sbin/make.cross: handle obj-$ARCH directory properly
If the directory obj-$ARCH exists, make.cross tries to use it as the
build directory. This happens even if you explicitly pass a different
directory to the O= option.

    $ mkdir obj-x86_64
    $ make.cross O=foo defconfig
    Compiler will be installed in /home/masahiro/0day
    make --keep-going KBUILD_OUTPUT=obj-x86_64 ARCH=x86_64 --jobs=32 O=foo defconfig
    make[1]: Entering directory '/home/masahiro/workspace/linux/obj-x86_64'
        [snip]
    make[1]: Leaving directory '/home/masahiro/workspace/linux/obj-x86_64'

The combination of 'KBUILD_OUTPUT=obj-x86_64' and 'O=foo' is odd.
Kbuild takes the former, and your request, O=foo, is just ignored.
obj-$ARCH should be considered only when the build directory is not
specified.

Another weirdness is, make.cross passes KBUILD_OUTPUT to the Make
command line, but KBUILD_OUTPUT should be an environment variable.
Pass O=obj-$ARCH instead of KBUILD_OUTPUT=obj-$ARCH to the command line.

Yet another bug is, BUILD_DIR is not set even if you have obj-$ARCH.
setup_compiler_for_* and setup_crosstool_* miss to check the .config
file in that directory.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>